+++ src/bot/handlers/index.ts
@@ -1,7 +1,7 @@
 export { default as logger } from "#handlers/logger";
 export { default as buildShardEmbed } from "#handlers/buildShardsEmbed";
 export * from "#handlers/getDailyEventTimes";
-export { default as getTS } from "#handlers/getTS";
+export { default as getTS } from "#bot/utils/getTSStatus";
 export { default as dblStats } from "#handlers/dblStats";
 export { default as topggAutopost } from "#handlers/topggAutopost";
 export * from "#handlers/sendDailyQuestReminder";
