+++ src/bot/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./getEventStatus.js";
+export { default as getTSStatus } from "./getTSStatus.js";
